Improve Mock Service Worker handlers folder structure
