+++ ui/src/mocks/handlers/constants.ts
@@ -1,0 +1,1 @@
+export const BASE_TEST_URL = 'http://localhost:8000';
