+++ ui/src/mocks/handlers/workspaces/responses.ts
@@ -1,0 +1,97 @@
+export const WORKSPACES_LIST = {
+  "content": [
+    {
+      "id": "da890485-b3a4-4ee3-96ca-290b06032af8",
+      "name": "                                                Sc",
+      "status": "INCOMPLETE",
+      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
+      "gitConfiguration": "",
+      "registryConfiguration": "",
+      "cdConfiguration": "",
+      "circleMatcherUrl": "",
+      "metricConfiguration": "",
+      "userGroups": [],
+      "createdAt": "2021-01-27 14:30:00"
+    },
+    {
+      "id": "7e41cd61-1d0a-4fca-9237-5b46705a6ed0",
+      "name": "                     sdfk.jhksjdfg                ",
+      "status": "INCOMPLETE",
+      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
+      "gitConfiguration": "",
+      "registryConfiguration": "",
+      "cdConfiguration": "",
+      "circleMatcherUrl": "",
+      "metricConfiguration": "",
+      "userGroups": [],
+      "createdAt": "2021-01-27 14:30:58"
+    },
+    {
+      "id": "cfc47018-c583-4e03-8c85-208334e68974",
+      "name": "12",
+      "status": "INCOMPLETE",
+      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
+      "gitConfiguration": "",
+      "registryConfiguration": "",
+      "cdConfiguration": "",
+      "circleMatcherUrl": "",
+      "metricConfiguration": "",
+      "userGroups": [],
+      "createdAt": "2021-01-26 12:54:26"
+    },
+    {
+      "id": "7ee6241e-9443-4c39-8b47-2d5bbd610c98",
+      "name": "Douglas",
+      "status": "INCOMPLETE",
+      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
+      "gitConfiguration": "",
+      "registryConfiguration": "",
+      "cdConfiguration": "",
+      "circleMatcherUrl": "",
+      "metricConfiguration": "",
+      "userGroups": [],
+      "createdAt": "2021-02-24 13:02:25"
+    },
+    {
+      "id": "f554b2e5-ba2f-45bb-afea-e239983c58dd",
+      "name": "Ieza tests",
+      "status": "COMPLETE",
+      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
+      "gitConfiguration": "",
+      "registryConfiguration": "",
+      "cdConfiguration": "",
+      "circleMatcherUrl": "",
+      "metricConfiguration": "",
+      "userGroups": [
+        {
+          "id": "58a706ed-06e8-4662-b52e-3c308087169c",
+          "name": " Profiling test",
+          "author": {
+            "id": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
+            "name": "rootqa",
+            "email": "rootqa@root",
+            "photoUrl": "",
+            "createdAt": "2020-11-27 13:30:16",
+            "root": false
+          },
+          "createdAt": "2020-11-30 18:17:48",
+          "users": [
+            {
+              "id": "a2e3a4fa-3e9e-47fa-949b-187c72f1b652",
+              "name": "Ieza Lopes",
+              "email": "ieza.damasceno@zup.com.br",
+              "photoUrl": "",
+              "createdAt": "2020-11-30 17:31:23",
+              "root": false
+            }
+          ]
+        }
+      ],
+      "createdAt": "2020-11-30 18:18:33"
+    }
+  ],
+  "page": 0,
+  "size": 5,
+  "totalPages": 3,
+  "last": false
+}
