--- ui/src/mocks/responses/workspaces.ts
@@ -1,1148 +1,0 @@
-export const WORKSPACES_LIST = {
-  "content": [
-    {
-      "id": "da890485-b3a4-4ee3-96ca-290b06032af8",
-      "name": "                                                Sc",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-27 14:30:00"
-    },
-    {
-      "id": "7e41cd61-1d0a-4fca-9237-5b46705a6ed0",
-      "name": "                     sdfk.jhksjdfg                ",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-27 14:30:58"
-    },
-    {
-      "id": "cfc47018-c583-4e03-8c85-208334e68974",
-      "name": "12",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-26 12:54:26"
-    },
-    {
-      "id": "7ee6241e-9443-4c39-8b47-2d5bbd610c98",
-      "name": "Douglas",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 13:02:25"
-    },
-    {
-      "id": "f554b2e5-ba2f-45bb-afea-e239983c58dd",
-      "name": "Ieza tests",
-      "status": "COMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "58a706ed-06e8-4662-b52e-3c308087169c",
-          "name": " Profiling test",
-          "author": {
-            "id": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-            "name": "rootqa",
-            "email": "rootqa@root",
-            "photoUrl": "",
-            "createdAt": "2020-11-27 13:30:16",
-            "root": false
-          },
-          "createdAt": "2020-11-30 18:17:48",
-          "users": [
-            {
-              "id": "a2e3a4fa-3e9e-47fa-949b-187c72f1b652",
-              "name": "Ieza Lopes",
-              "email": "ieza.damasceno@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-30 17:31:23",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2020-11-30 18:18:33"
-    },
-    {
-      "id": "1f57eb49-2993-4478-8fd2-00eb42fd83d2",
-      "name": "Metrics - teste 00",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-08 13:11:26"
-    },
-    {
-      "id": "05ddaf5b-a4e0-4709-8d18-445fd7a966fb",
-      "name": "Sales treta 1",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-01 20:45:43"
-    },
-    {
-      "id": "6445ffb7-eb46-4068-a089-70386f8eb988",
-      "name": "Sales treta 2",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-01 20:48:29"
-    },
-    {
-      "id": "b8a7a386-cf27-4e64-88cf-25a1424f6af7",
-      "name": "Schu",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-26 12:55:13"
-    },
-    {
-      "id": "6b06daab-2c8f-4c76-9c97-452aa9f1abec",
-      "name": "Schu                     1234",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-26 14:06:59"
-    },
-    {
-      "id": "6f356bad-8d14-4ddc-9c52-2d6c063f0f83",
-      "name": "Schu 1234",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-28 17:35:30"
-    },
-    {
-      "id": "2daa2e98-1045-411f-a18f-1f0673af36a5",
-      "name": "Squad-Deploy",
-      "status": "COMPLETE",
-      "authorId": "adf716ec-4724-4f5e-88ac-eabbb44cf1ff",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "9ca41b9e-889e-4e51-acc3-61dc455f7f30",
-          "name": "Maintenance",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2020-11-27 18:36:34",
-          "users": [
-            {
-              "id": "4bbb5220-3b37-4847-8467-2a6c889f5a3c",
-              "name": "Bárbara Rossalli",
-              "email": "barbara.rocha@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 18:45:08",
-              "root": false
-            },
-            {
-              "id": "ab2b89a9-53c1-462d-90a2-7cf1dab9c3f0",
-              "name": "Douglas",
-              "email": "douglas.fernandes@zup.com.br",
-              "photoUrl": "https://charlescd.io",
-              "createdAt": "2020-11-20 18:52:04",
-              "root": false
-            },
-            {
-              "id": "11fd1131-2eef-4266-911d-30b697f92769",
-              "name": "Maintenance",
-              "email": "maintenance@zup.com",
-              "photoUrl": "",
-              "createdAt": "2020-11-27 18:32:30",
-              "root": false
-            },
-            {
-              "id": "a3f3a41a-1b23-4326-9014-3b08fbb351ae",
-              "name": "Felipe Brunelli de Andrade",
-              "email": "felipe.andrade@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-14 20:27:11",
-              "root": false
-            },
-            {
-              "id": "67c389b4-873c-47d2-82fd-c1f55d4a8b1a",
-              "name": "angelica",
-              "email": "angelica.lima@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-20 13:24:36",
-              "root": false
-            },
-            {
-              "id": "20960e66-a619-4f05-b9f3-f0361e40c944",
-              "name": "Leandro",
-              "email": "leandro.queiroz@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 21:00:32",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "00572c2c-7be3-4c6c-99b9-a4aa5bde0193",
-          "name": "Developers Of Heaven",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2021-01-27 18:41:56",
-          "users": [
-            {
-              "id": "4bbb5220-3b37-4847-8467-2a6c889f5a3c",
-              "name": "Bárbara Rossalli",
-              "email": "barbara.rocha@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 18:45:08",
-              "root": false
-            },
-            {
-              "id": "ab2b89a9-53c1-462d-90a2-7cf1dab9c3f0",
-              "name": "Douglas",
-              "email": "douglas.fernandes@zup.com.br",
-              "photoUrl": "https://charlescd.io",
-              "createdAt": "2020-11-20 18:52:04",
-              "root": false
-            },
-            {
-              "id": "79c358a8-4d33-48ec-ba42-1d9b9f0d9220",
-              "name": "Diplodocus DevTester",
-              "email": "diplodocus@dino.com",
-              "photoUrl": "",
-              "createdAt": "2021-02-08 18:28:50",
-              "root": false
-            },
-            {
-              "id": "41e19170-0fe9-46cc-a86b-e2e9aedfa107",
-              "name": "Pedro Borges",
-              "email": "pedro.borges@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-02 04:11:34",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2020-11-20 13:46:43"
-    },
-    {
-      "id": "be83b2ef-0fa7-40a7-a9cd-87d6a7ca18ab",
-      "name": "Squad-Devx",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "https://charles-sandbox.continuousplatform.com/charlescd-circle-matcher",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "58a706ed-06e8-4662-b52e-3c308087169c",
-          "name": " Profiling test",
-          "author": {
-            "id": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-            "name": "rootqa",
-            "email": "rootqa@root",
-            "photoUrl": "",
-            "createdAt": "2020-11-27 13:30:16",
-            "root": false
-          },
-          "createdAt": "2020-11-30 18:17:48",
-          "users": [
-            {
-              "id": "a2e3a4fa-3e9e-47fa-949b-187c72f1b652",
-              "name": "Ieza Lopes",
-              "email": "ieza.damasceno@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-30 17:31:23",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "00572c2c-7be3-4c6c-99b9-a4aa5bde0193",
-          "name": "Developers Of Heaven",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2021-01-27 18:41:56",
-          "users": [
-            {
-              "id": "4bbb5220-3b37-4847-8467-2a6c889f5a3c",
-              "name": "Bárbara Rossalli",
-              "email": "barbara.rocha@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 18:45:08",
-              "root": false
-            },
-            {
-              "id": "ab2b89a9-53c1-462d-90a2-7cf1dab9c3f0",
-              "name": "Douglas",
-              "email": "douglas.fernandes@zup.com.br",
-              "photoUrl": "https://charlescd.io",
-              "createdAt": "2020-11-20 18:52:04",
-              "root": false
-            },
-            {
-              "id": "79c358a8-4d33-48ec-ba42-1d9b9f0d9220",
-              "name": "Diplodocus DevTester",
-              "email": "diplodocus@dino.com",
-              "photoUrl": "",
-              "createdAt": "2021-02-08 18:28:50",
-              "root": false
-            },
-            {
-              "id": "41e19170-0fe9-46cc-a86b-e2e9aedfa107",
-              "name": "Pedro Borges",
-              "email": "pedro.borges@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-02 04:11:34",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "9d6398c0-8752-4291-831f-8b21be88ea31",
-          "name": "DEVX",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2021-02-04 12:43:19",
-          "users": [
-            {
-              "id": "956b138f-df17-4292-b0bc-7e0b65f9fa6c",
-              "name": "luane (gmail)",
-              "email": "luane@gmail.com",
-              "photoUrl": "",
-              "createdAt": "2020-12-03 18:48:13",
-              "root": false
-            },
-            {
-              "id": "79c358a8-4d33-48ec-ba42-1d9b9f0d9220",
-              "name": "Diplodocus DevTester",
-              "email": "diplodocus@dino.com",
-              "photoUrl": "",
-              "createdAt": "2021-02-08 18:28:50",
-              "root": false
-            },
-            {
-              "id": "b421820b-fd00-4a3f-be47-9cca5a05fe96",
-              "name": "Monica Ribeiro",
-              "email": "monica.ribeiro@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-30 20:43:08",
-              "root": false
-            },
-            {
-              "id": "41e19170-0fe9-46cc-a86b-e2e9aedfa107",
-              "name": "Pedro Borges",
-              "email": "pedro.borges@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-02 04:11:34",
-              "root": false
-            },
-            {
-              "id": "4bbb5220-3b37-4847-8467-2a6c889f5a3c",
-              "name": "Bárbara Rossalli",
-              "email": "barbara.rocha@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 18:45:08",
-              "root": false
-            },
-            {
-              "id": "8568c0bc-60cb-4f45-9599-98c1367d025a",
-              "name": "schumaiquer",
-              "email": "schumaiquer@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-20 13:26:17",
-              "root": false
-            },
-            {
-              "id": "20960e66-a619-4f05-b9f3-f0361e40c944",
-              "name": "Leandro",
-              "email": "leandro.queiroz@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 21:00:32",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2020-12-02 17:04:52"
-    },
-    {
-      "id": "944c66f7-5583-4273-87a2-e1fd2206ce5f",
-      "name": "Squad-Metrics",
-      "status": "COMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "http://charlescd-circle-matcher:8080",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "9ca41b9e-889e-4e51-acc3-61dc455f7f30",
-          "name": "Maintenance",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2020-11-27 18:36:34",
-          "users": [
-            {
-              "id": "4bbb5220-3b37-4847-8467-2a6c889f5a3c",
-              "name": "Bárbara Rossalli",
-              "email": "barbara.rocha@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 18:45:08",
-              "root": false
-            },
-            {
-              "id": "ab2b89a9-53c1-462d-90a2-7cf1dab9c3f0",
-              "name": "Douglas",
-              "email": "douglas.fernandes@zup.com.br",
-              "photoUrl": "https://charlescd.io",
-              "createdAt": "2020-11-20 18:52:04",
-              "root": false
-            },
-            {
-              "id": "11fd1131-2eef-4266-911d-30b697f92769",
-              "name": "Maintenance",
-              "email": "maintenance@zup.com",
-              "photoUrl": "",
-              "createdAt": "2020-11-27 18:32:30",
-              "root": false
-            },
-            {
-              "id": "a3f3a41a-1b23-4326-9014-3b08fbb351ae",
-              "name": "Felipe Brunelli de Andrade",
-              "email": "felipe.andrade@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-14 20:27:11",
-              "root": false
-            },
-            {
-              "id": "67c389b4-873c-47d2-82fd-c1f55d4a8b1a",
-              "name": "angelica",
-              "email": "angelica.lima@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-20 13:24:36",
-              "root": false
-            },
-            {
-              "id": "20960e66-a619-4f05-b9f3-f0361e40c944",
-              "name": "Leandro",
-              "email": "leandro.queiroz@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-12-08 21:00:32",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "a641d8b1-17a4-4c85-8b65-9be0099f4271",
-          "name": "Write",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2020-11-27 18:38:36",
-          "users": [
-            {
-              "id": "ff33b609-3dbd-47dc-8762-43c56ea74b54",
-              "name": "Write",
-              "email": "write@zup.com",
-              "photoUrl": "",
-              "createdAt": "2020-11-27 18:33:08",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "dae1a55a-c075-4d04-9e79-fcc02e68a7ee",
-          "name": "Metrics-Group",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2020-11-20 13:49:30",
-          "users": [
-            {
-              "id": "4ee48b6d-613e-47de-9d4d-9a47657c1dfc",
-              "name": "icaro",
-              "email": "icaro.afonso@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-20 13:32:56",
-              "root": false
-            },
-            {
-              "id": "bc4bd99b-3575-4e05-98cf-8d77b0ade5ea",
-              "name": "Schu",
-              "email": "schu@zup.com",
-              "photoUrl": "",
-              "createdAt": "2021-01-21 14:18:34",
-              "root": false
-            },
-            {
-              "id": "d6a9e1a7-4932-44c6-a038-f888ab2ab88a",
-              "name": "Catanant",
-              "email": "bruno.catanant@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2021-01-12 19:26:50",
-              "root": false
-            },
-            {
-              "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-              "name": "Charles Adminn",
-              "email": "charlesadmin@admin",
-              "photoUrl": "",
-              "createdAt": "2020-11-16 18:20:57",
-              "root": false
-            },
-            {
-              "id": "8568c0bc-60cb-4f45-9599-98c1367d025a",
-              "name": "schumaiquer",
-              "email": "schumaiquer@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2020-11-20 13:26:17",
-              "root": false
-            }
-          ]
-        },
-        {
-          "id": "38963327-ace9-47c8-a0fd-3100a9da3896",
-          "name": "Read",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2020-11-27 18:38:15",
-          "users": [
-            {
-              "id": "4ed0a11c-ad7f-4e2c-b3fe-cb3217875998",
-              "name": "Read",
-              "email": "read@zup.com",
-              "photoUrl": "",
-              "createdAt": "2020-11-27 18:33:35",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2020-11-20 13:44:27"
-    },
-    {
-      "id": "4240c632-61c1-4b17-9df8-7b5f3de33f0f",
-      "name": "Test automation 5e4bd56b09",
-      "status": "COMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "http://charlescd-circle-matcher:8080",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "8fb3aeac-0c02-48dd-837b-51f5d2eea811",
-          "name": "Test automation 5e4bd56b09",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2021-01-13 12:04:18",
-          "users": [
-            {
-              "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-              "name": "Charles Adminn",
-              "email": "charlesadmin@admin",
-              "photoUrl": "",
-              "createdAt": "2020-11-16 18:20:57",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2021-01-13 12:04:27"
-    },
-    {
-      "id": "ff8a5f5f-2c9a-4eaa-a417-bba6b3b6e42e",
-      "name": "asd",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-26 13:03:12"
-    },
-    {
-      "id": "a2153af8-222a-45d1-ab16-874a01bab7e8",
-      "name": "dasdasdas",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-29 18:28:41"
-    },
-    {
-      "id": "894a68dd-f2ef-4a27-b6e4-dd4a52ab0404",
-      "name": "deleteme",
-      "status": "COMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-09 17:43:48"
-    },
-    {
-      "id": "0ca62c37-9fe0-46a3-b9ac-3920b979b4c0",
-      "name": "deleteme 2",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-09 17:45:16"
-    },
-    {
-      "id": "9e9b9f7b-dbbd-4ccc-b72f-01cff53c7faa",
-      "name": "deleteme 3",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-09 18:04:20"
-    },
-    {
-      "id": "3cb0effc-8776-4643-9a6a-0ec55ab82da2",
-      "name": "dssdfsd",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-08 13:17:58"
-    },
-    {
-      "id": "a1ccc5f9-694e-49f5-8cbf-66d9ca260572",
-      "name": "name",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-21 20:13:11"
-    },
-    {
-      "id": "10da353d-4392-469a-8084-27d752187b9a",
-      "name": "percentage-sls",
-      "status": "COMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "http://charlescd-circle-matcher:8080",
-      "metricConfiguration": "",
-      "userGroups": [
-        {
-          "id": "822bcb76-8bbb-4c02-abad-7a9cfca7e254",
-          "name": "percentage-sls",
-          "author": {
-            "id": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-            "name": "Charles Adminn",
-            "email": "charlesadmin@admin",
-            "photoUrl": "",
-            "createdAt": "2020-11-16 18:20:57",
-            "root": false
-          },
-          "createdAt": "2021-01-29 12:40:04",
-          "users": [
-            {
-              "id": "298bbca2-414d-4a34-9043-3dc4a7ba1ca1",
-              "name": "lucas.sales@zup.com.br",
-              "email": "lucas.sales@zup.com.br",
-              "photoUrl": "",
-              "createdAt": "2021-01-29 12:37:19",
-              "root": false
-            }
-          ]
-        }
-      ],
-      "createdAt": "2021-01-29 12:37:39"
-    },
-    {
-      "id": "521e96cd-70b8-4288-9c93-8f797fa8be1c",
-      "name": "pineapple",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-25 18:51:53"
-    },
-    {
-      "id": "80070134-a78c-424b-9655-5e1422428dfe",
-      "name": "sales-test",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "http://charlescd-circle-matcher:8080",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2020-12-10 17:15:36"
-    },
-    {
-      "id": "1ef03526-3bf3-406a-a943-91e983b77e04",
-      "name": "sales-vivo-test",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-01-27 12:44:51"
-    },
-    {
-      "id": "ada391fb-382a-4513-8c3a-fa03461efac8",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:45"
-    },
-    {
-      "id": "0992182a-b96e-488f-8af2-ceb65c811bb6",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:52:08"
-    },
-    {
-      "id": "d1c719b2-97a0-4f55-804f-3aef402076bf",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:52:14"
-    },
-    {
-      "id": "d35ab129-7de7-4c30-aea5-f784d8868d7f",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:46"
-    },
-    {
-      "id": "e16fe615-4089-47ac-a1d4-4ff2ec293e7b",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:49"
-    },
-    {
-      "id": "5189a511-d7c6-4b49-bfb9-ee33628b164c",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:49:42"
-    },
-    {
-      "id": "147a19d2-3577-46c1-9ad7-13c32b80df86",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:54"
-    },
-    {
-      "id": "da87e88a-c78d-4ddb-9690-c9aec45d70cc",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:51"
-    },
-    {
-      "id": "acab833d-17ea-457e-acc2-fb8a081ca3ec",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:50"
-    },
-    {
-      "id": "ecf17298-33c8-4a81-a01f-0f6c1fcfbbda",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:52"
-    },
-    {
-      "id": "78f339c0-7da1-4629-aa8e-51b58f1ad08c",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:48"
-    },
-    {
-      "id": "8f6c062f-2bdd-49f1-9ef1-b463d99fd6b7",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:49:26"
-    },
-    {
-      "id": "f7208d69-4488-43b4-adca-557162f1fd26",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:55"
-    },
-    {
-      "id": "9c4e18c7-ae75-4faf-878a-6e5ef7170d87",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:52:08"
-    },
-    {
-      "id": "97914b53-b69d-4ecd-9659-459db22b1667",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:47"
-    },
-    {
-      "id": "07e95026-5ee4-49b5-8f1c-9c863a494c48",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:53"
-    },
-    {
-      "id": "648b561e-2241-4223-b855-acce33fd1a29",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:51"
-    },
-    {
-      "id": "4c3b745b-8c79-4520-afb6-e99e3ae00003",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:52:07"
-    },
-    {
-      "id": "5a447540-deb4-4628-8450-58ad1601cac1",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:46"
-    },
-    {
-      "id": "abeacf47-7960-4cec-9881-cae34015163f",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:54"
-    },
-    {
-      "id": "c00dfb99-6015-4ec9-ae4b-cd0994d9d163",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:49"
-    },
-    {
-      "id": "9d162005-ab0b-4e37-88f1-c58e16ee4a23",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:53"
-    },
-    {
-      "id": "2c243a56-e036-41d5-ac40-912fefed0491",
-      "name": "test create",
-      "status": "INCOMPLETE",
-      "authorId": "1770c44f-6c5f-4cac-b29d-72fd3384b5f7",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-24 17:51:44"
-    },
-    {
-      "id": "28f30ed8-743b-4455-b44a-d7b4d46eadfd",
-      "name": "test-active",
-      "status": "INCOMPLETE",
-      "authorId": "c7e6dafe-aa7a-4536-be1b-34eaad4c2915",
-      "gitConfiguration": "",
-      "registryConfiguration": "",
-      "cdConfiguration": "",
-      "circleMatcherUrl": "",
-      "metricConfiguration": "",
-      "userGroups": [],
-      "createdAt": "2021-02-10 20:53:36"
-    }
-  ],
-  "page": 0,
-  "size": 50,
-  "totalPages": 3,
-  "last": false
-}
